Move NounPhrase to NaturalLanguage.English
